+++ backend/src/main/java/develup/api/DiscussionApi.java
@@ -5,8 +5,9 @@ import develup.api.auth.Auth;
 import develup.api.common.ApiResponse;
 import develup.application.auth.Accessor;
 import develup.application.discussion.CreateDiscussionRequest;
+import develup.application.discussion.DiscussionReadService;
 import develup.application.discussion.DiscussionResponse;
-import develup.application.discussion.DiscussionService;
+import develup.application.discussion.DiscussionWriteService;
 import develup.application.discussion.SummarizedDiscussionResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -23,10 +24,12 @@ import org.springframework.web.bind.annotation.RestController;
 @Tag(name = "디스커션 API")
 public class DiscussionApi {
 
-    private final DiscussionService discussionService;
+    private final DiscussionWriteService discussionWriteService;
+    private final DiscussionReadService discussionReadService;
 
-    public DiscussionApi(DiscussionService discussionService) {
-        this.discussionService = discussionService;
+    public DiscussionApi(DiscussionWriteService discussionWriteService, DiscussionReadService discussionReadService) {
+        this.discussionWriteService = discussionWriteService;
+        this.discussionReadService = discussionReadService;
     }
 
     @GetMapping("/discussions")
@@ -35,7 +38,7 @@ public class DiscussionApi {
             @RequestParam(defaultValue = "all") String mission,
             @RequestParam(defaultValue = "all") String hashTag
     ) {
-        List<SummarizedDiscussionResponse> responses = discussionService.getSummaries(mission, hashTag);
+        List<SummarizedDiscussionResponse> responses = discussionReadService.getSummaries(mission, hashTag);
 
         return ResponseEntity.ok(new ApiResponse<>(responses));
     }
@@ -43,7 +46,7 @@ public class DiscussionApi {
     @GetMapping("/discussions/{id}")
     @Operation(summary = "디스커션 조회 API", description = "디스커션을 조회합니다.")
     public ResponseEntity<ApiResponse<DiscussionResponse>> getDiscussion(@PathVariable Long id) {
-        DiscussionResponse response = discussionService.getById(id);
+        DiscussionResponse response = discussionReadService.getById(id);
 
         return ResponseEntity.ok(new ApiResponse<>(response));
     }
@@ -54,7 +57,7 @@ public class DiscussionApi {
             @Auth Accessor accessor,
             @Valid @RequestBody CreateDiscussionRequest request
     ) {
-        DiscussionResponse response = discussionService.create(accessor.id(), request);
+        DiscussionResponse response = discussionWriteService.create(accessor.id(), request);
 
         return ResponseEntity.ok(new ApiResponse<>(response));
     }
@@ -62,7 +65,7 @@ public class DiscussionApi {
     @GetMapping("/discussions/mine")
     @Operation(summary = "나의 디스커션 목록 조회 API", description = "내가 작성한 디스커션 목록을 조회합니다.")
     public ResponseEntity<ApiResponse<List<SummarizedDiscussionResponse>>> getMyDiscussions(@Auth Accessor accessor) {
-        List<SummarizedDiscussionResponse> response = discussionService.getDiscussionsByMemberId(accessor.id());
+        List<SummarizedDiscussionResponse> response = discussionReadService.getDiscussionsByMemberId(accessor.id());
 
         return ResponseEntity.ok(new ApiResponse<>(response));
     }
