+++ backend/src/test/java/develup/application/discussion/DiscussionWriteServiceTest.java
@@ -8,31 +8,21 @@ import static org.junit.jupiter.api.Assertions.assertNull;
 
 import java.util.List;
 import develup.api.exception.DevelupException;
-import develup.domain.discussion.Discussion;
-import develup.domain.discussion.DiscussionRepository;
-import develup.domain.hashtag.HashTag;
 import develup.domain.hashtag.HashTagRepository;
-import develup.domain.member.Member;
 import develup.domain.member.MemberRepository;
-import develup.domain.mission.Mission;
 import develup.domain.mission.MissionRepository;
 import develup.support.IntegrationTestSupport;
-import develup.support.data.DiscussionTestData;
 import develup.support.data.HashTagTestData;
 import develup.support.data.MemberTestData;
 import develup.support.data.MissionTestData;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.transaction.annotation.Transactional;
 
-class DiscussionServiceTest extends IntegrationTestSupport {
-
-    @Autowired
-    private DiscussionService discussionService;
+class DiscussionWriteServiceTest extends IntegrationTestSupport {
 
     @Autowired
-    private DiscussionRepository discussionRepository;
+    private DiscussionWriteService discussionWriteService;
 
     @Autowired
     private MemberRepository memberRepository;
@@ -44,26 +34,6 @@ class DiscussionServiceTest extends IntegrationTestSupport {
     private HashTagRepository hashTagRepository;
 
     @Test
-    @DisplayName("디스커션 리스트를 조회한다.")
-    @Transactional
-    void getSummaries() {
-        Mission mission1 = missionRepository.save(MissionTestData.defaultMission().withTitle("루터회관 흡연단속").build());
-        Mission mission2 = missionRepository.save(MissionTestData.defaultMission().withTitle("주문").build());
-        HashTag hashTag1 = hashTagRepository.save(HashTagTestData.defaultHashTag().withName("JAVA").build());
-        HashTag hashTag2 = hashTagRepository.save(HashTagTestData.defaultHashTag().withName("객체지향").build());
-        createDiscussion(mission1, hashTag1);
-        createDiscussion(mission2, hashTag2);
-
-        assertAll(
-                () -> assertThat(discussionService.getSummaries("all", "all")).hasSize(2),
-                () -> assertThat(discussionService.getSummaries(mission1.getTitle(), "all")).hasSize(1),
-                () -> assertThat(discussionService.getSummaries(mission2.getTitle(), "all")).hasSize(1),
-                () -> assertThat(discussionService.getSummaries("all", hashTag1.getName())).hasSize(1),
-                () -> assertThat(discussionService.getSummaries("all", hashTag2.getName())).hasSize(1)
-        );
-    }
-
-    @Test
     @DisplayName("디스커션을 제출한다.")
     void create() {
         Long memberId = memberRepository.save(MemberTestData.defaultMember().build()).getId();
@@ -73,7 +43,7 @@ class DiscussionServiceTest extends IntegrationTestSupport {
         CreateDiscussionRequest request =
                 new CreateDiscussionRequest("title", "content", missionId, List.of(hashTagId));
 
-        DiscussionResponse response = discussionService.create(memberId, request);
+        DiscussionResponse response = discussionWriteService.create(memberId, request);
 
         assertAll(
                 () -> assertEquals(response.id(), 1L),
@@ -93,7 +63,7 @@ class DiscussionServiceTest extends IntegrationTestSupport {
         CreateDiscussionRequest request =
                 new CreateDiscussionRequest("title", "content", null, List.of(hashTagId));
 
-        DiscussionResponse response = discussionService.create(memberId, request);
+        DiscussionResponse response = discussionWriteService.create(memberId, request);
 
         assertAll(
                 () -> assertEquals(response.id(), 1L),
@@ -113,7 +83,7 @@ class DiscussionServiceTest extends IntegrationTestSupport {
         CreateDiscussionRequest request =
                 new CreateDiscussionRequest("title", "content", missionId, List.of());
 
-        DiscussionResponse response = discussionService.create(memberId, request);
+        DiscussionResponse response = discussionWriteService.create(memberId, request);
 
         assertAll(
                 () -> assertEquals(response.id(), 1L),
@@ -134,7 +104,7 @@ class DiscussionServiceTest extends IntegrationTestSupport {
         CreateDiscussionRequest request =
                 new CreateDiscussionRequest("title", "content", missionId, List.of(hashTagId));
 
-        assertThatThrownBy(() -> discussionService.create(unknownMemberId, request))
+        assertThatThrownBy(() -> discussionWriteService.create(unknownMemberId, request))
                 .isInstanceOf(DevelupException.class)
                 .hasMessage("존재하지 않는 회원입니다.");
     }
@@ -149,7 +119,7 @@ class DiscussionServiceTest extends IntegrationTestSupport {
         CreateDiscussionRequest request =
                 new CreateDiscussionRequest("title", "content", unknownMissionId, List.of(hashTagId));
 
-        assertThatThrownBy(() -> discussionService.create(memberId, request))
+        assertThatThrownBy(() -> discussionWriteService.create(memberId, request))
                 .isInstanceOf(DevelupException.class)
                 .hasMessage("존재하지 않는 미션입니다.");
     }
@@ -164,48 +134,8 @@ class DiscussionServiceTest extends IntegrationTestSupport {
         CreateDiscussionRequest request =
                 new CreateDiscussionRequest("title", "content", missionId, List.of(unknownHashTagId));
 
-        assertThatThrownBy(() -> discussionService.create(memberId, request))
+        assertThatThrownBy(() -> discussionWriteService.create(memberId, request))
                 .isInstanceOf(DevelupException.class)
                 .hasMessage("존재하지 않는 해시태그입니다.");
-    }
-
-    @Test
-    @DisplayName("존재하지 않는 디스커션은 불러올 수 없다.")
-    void getById() {
-        Long unknownId = -1L;
-
-        assertThatThrownBy(() -> discussionService.getById(unknownId))
-                .isInstanceOf(DevelupException.class)
-                .hasMessage("존재하지 않는 디스커션입니다.");
-    }
-
-    @Test
-    @DisplayName("나의 디스커션 리스트를 조회한다.")
-    @Transactional
-    void getDiscussionsByMemberId() {
-        Member member = memberRepository.save(MemberTestData.defaultMember().withId(1L).build());
-        Mission mission = missionRepository.save(MissionTestData.defaultMission().build());
-        HashTag hashTag = hashTagRepository.save(HashTagTestData.defaultHashTag().build());
-        Discussion discussion = DiscussionTestData.defaultDiscussion()
-                .withMember(member)
-                .withMission(mission)
-                .withHashTags(List.of(hashTag))
-                .build();
-
-        discussionRepository.save(discussion);
-
-        assertThat(discussionService.getDiscussionsByMemberId(member.getId())).hasSize(1);
-    }
-
-    private void createDiscussion(Mission mission, HashTag hashTag) {
-        Member member = memberRepository.save(MemberTestData.defaultMember().build());
-
-        Discussion discussion = DiscussionTestData.defaultDiscussion()
-                .withMission(mission)
-                .withMember(member)
-                .withHashTags(List.of(hashTag))
-                .build();
-
-        discussionRepository.save(discussion);
     }
 }
